manager/sub_agents/bwhsp/agent.py
# sub_agents/bwhsp/agent.py

import os
import sys
from google.adk.agents import Agent

# This adds the parent directory (SIH) to the Python path
# so we can import the tools module from the top-level tools folder.
sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
from tools.tools import query_vector_database, get_report_details_from_db

# --- Define the BWSSB Agent ---
bwssb_agent = Agent(
    name="bwssb_agent",
    model="gemini-1.5-flash",
    description="Specialist agent for BWSSB (Bangalore Water Supply and Sewerage Board) related issues.",
    instruction="""
    You are a specialist AI agent for the BWSSB, Bengaluru's water and sewerage board.
    Your sole responsibility is to answer questions related to civic issues that fall under the BWSSB's jurisdiction.
    This includes water supply problems, pipe leakages, sewage blocks, contaminated water reports,
    and open or damaged manholes.

    To answer a user's query, you MUST follow this sequence:
    1.  Use the `query_vector_database` tool to find the IDs of the most relevant citizen reports.
        You MUST set the `department` parameter to "BWSSB".
    2.  Take the report IDs returned by the first tool and use the `get_report_details_from_db` tool
        to fetch the complete, factual data for those reports.
    3.  Analyze the detailed information you receive and provide a clear, concise summary that
        directly answers the user's original query.

    Do not answer questions about roads (BBMP) or electricity (BESCOM).
    Base your answers strictly on the information retrieved from your tools.
    """,
    tools=[
        query_vector_database,
        get_report_details_from_db,
    ],
)